Rename header files to use prefix mh_ to clearly indicate which headers are mugy's. Rename a couple of files to remove _mugy since its redundant.
