+++ include/mh_finalize.h
@@ -6,8 +6,8 @@
 #ifndef FINALIZE
 #define FINALIZE
 
-#include "io_tools.h"
-#include "mpi_tools.h"
+#include "mh_io_tools.h"
+#include "mh_mpi_tools.h"
 
 // Call for the termination of the various parts of mugy.
 void terminate_all();
