+++ include/mh_initialization.h
@@ -6,13 +6,13 @@
 #ifndef INITIALIZATION
 #define INITIALIZATION
 
-#include "parameters.h"
-#include "data_mugy.h"
-#include "utilities.h"
-#include "alloc_mugy.h"
 #include <string.h>   // e.g. for strcat, strlen.
-#include "mpi_tools.h"
-#include "io_tools.h"
+#include "mh_parameters.h"
+#include "mh_data.h"
+#include "mh_utilities.h"
+#include "mh_alloc.h"
+#include "mh_mpi_tools.h"
+#include "mh_io_tools.h"
 
 // Read a real variable from input file. Need this function to support
 // reading floats and doubles with the same function call.
