+++ include/mh_io_adios.h
@@ -7,9 +7,9 @@
 #define IO_ADIOS
 
 #include "adios2_c.h"
-#include "mugyFLAGS.h"
-#include "data_mugy.h"
-#include "utilities.h"
+#include "mh_userFLAGS.h"
+#include "mh_data.h"
+#include "mh_utilities.h"
 
 extern adios2_adios *ad_ctx;  // ADIOS context used throughout our IO.
