+++ include/mh_mugy.h
@@ -6,28 +6,28 @@
 */
 
 // Pre-processor flags.
-#include "mugyFLAGS.h"
+#include "mh_userFLAGS.h"
 
 // Global parameters.
-#include "parameters.h"
+#include "mh_parameters.h"
 
 // Generic utility functions (with few or no dependencies).
-#include "utilities.h"
+#include "mh_utilities.h"
 
 // MPI-related infrastructure.
-#include "mpi_tools.h"
+#include "mh_mpi_tools.h"
 
 // IO module.
-#include "io_tools.h"
+#include "mh_io_tools.h"
 
 // Data structures specific to mugy.
-#include "data_mugy.h"
+#include "mh_data.h"
 
 // Functions that allocate mugy data.
-#include "alloc_mugy.h"
+#include "mh_alloc.h"
 
 // Functions used to initialize the simulation.
-#include "initialization.h"
+#include "mh_initialization.h"
 
 // Module in charge of closing mugy sim.
-#include "finalize.h"
+#include "mh_finalize.h"
