+++ include/mh_parameters.h
@@ -6,7 +6,7 @@
 #ifndef PARAMETERS
 #define PARAMETERS
 
-#include "data_mugy.h"
+#include "mh_data.h"
 
 extern mint nFrames, mpiProcs[nDim];
 extern real kxMaxDyn[nDim], dt, endTime;
