+++ src/mugy.in
@@ -21,7 +21,7 @@ ark_ewtScaling   = 0
 
 &species
 numSpecies = 2
-mpiProcs   = 1
+mpiProcs   = 2
 numMoments = 2 2
 qCharge    = -3.0 4.0
 muMass     = 1.0 60.59
